% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CITMIC_Data}
\alias{CITMIC_Data}
\title{An environment variable that includes some example data}
\format{
An environment variable
}
\usage{
CITMIC_Data
}
\description{
matirx_cell_go_inter:A matrix of Jaccard score between cells and GOBP.
matirx_cell_go_jaccard:A matrix consisting of genes shared by cells targets and GOBP.
GEP:An example gene expression profile.
}
\keyword{datasets}
