% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData_CITMIC.R
\name{GetData_CITMIC}
\alias{GetData_CITMIC}
\title{GetData_CITMIC}
\usage{
GetData_CITMIC(Data)
}
\arguments{
\item{Data}{A character should be one of "GEP", "matrix_cell_go_inter", "matrix_cell_go_jaccard"}
}
\value{
Data
}
\description{
Get the example data
}
