% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITMIC.R
\name{CITMIC}
\alias{CITMIC}
\title{CITMIC}
\usage{
CITMIC(GEP, damping=0.90)
}
\arguments{
\item{GEP}{An example gene expression profile}

\item{damping}{Restart the probability of the random-walk algorithm (default: 0.9).}
}
\value{
cell infiltration score
}
\description{
The function "CITMIC" is used to identify cell infiltration in tumor microenvironment by calculating intercellular crosstalk.
}
\examples{
# Obtain the example data
GEP<-GetData_CITMIC("GEP")
# Run the function
\donttest{lnScore<-CITMIC(GEP, damping=0.90)}
}
